move HasClearButton to component package
