+++ vaadin-flow-components-shared-parent/vaadin-flow-components-base/src/main/java/com/vaadin/flow/component/HasClearButton.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.mixin;
+package com.vaadin.flow.component;
 
 import com.vaadin.flow.component.HasElement;
